dynamic interior node bounds and centroids
While computing the Barnes-Hut solution, the tree is now rebalanced
after each iteration. When a node is inserted the resulting change in
all of the parent interior nodes is propagated up the tree recursively.

Centroids are now computed as the center of mass of the content of an
interior node rather than simply the spatial center of that node. This
prevents pathological insertion cases of nodes that are very close
together that can can create an asymptotically infinite number of splits
as leaf nodes get infinitesimally close together.